guide: iOS Onion Browser leaking IP
Since 2018, it is known to have a leak in `RTCPeerConnection`.
This is something that the developers have said cannot be
removed and thus, the functionality of the browser still
contains this call, and users should be wary of using iOS
until it is fixed. We will monitor the situation but it doesn't
appear they can patch.

Signed-off-by: pterocles <pterocles@proton.me>