Add link to Arkenfox/user.js
Arkenfox relies on a set of `about:config` tweaks in
Firefox. It allows you to set many privacy and disk 
avoidance related values which provide you with the 
ability to avoid useless Firefox additions and also 
avoid giving your browser a unique value which can
be tracked across the web.

Signed-off-by: pterocles <pterocles@proton.me>