Proposing an update to the about page
Proposing an addition to make sure readers don't think it's a solo project but is now an organization with multiple projects, collaborators, and looking for more collaborators. The text without this addition make it look like it's still a solo project and might deter/confuse people from asking to collaborate on the project as it gives an impression of being a solo project (to me at least).