Doc: Microarchitecture bugs in Intel and AMD CPUs
Add information about microarchitecture bugs

It's been a very busy year for Intel and AMD CPUs. Researchers have
discovered bugs left and right. One of them was revealed recently and I
felt compelled to add not only it but also information regarding other
types of attacks that aren't yet in the guide.

Signed-off-by: pterocles <pterocles@proton.me>