Revisiting and updating two sections
* Your Browser and Device Fingerprints:
* Microarchitectural Side-channel Deanonymization Attacks

Adds previous documentation on side-channel attacks which can be used
to fingerprint users, presented in 2016. This means this type of
microarchitectural deanonymization attack has been performed
possibly as early on as the first day websites have been able to
perform browser fingerprinting.

Fixes #70: ("Revisiting the Browser and Device Fingerprints &
Microarchitectural Side-channel Deanonymization Attacks sections")

We have two sections for FPing and this clearly defines the meaning of both:
- one section is a brief explanation and intro into Fingerprinting and "why"
- the other provides more technical analysis of things that can be leaked

Signed-off-by: Sharp-tailed Grouse <sharptail@riseup.net>